requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
langchain
langchain-google-genai
google-generativeai
